Updated hardware/firmware that adds an optional reset header
The optional reset header that can be used to hookup openkey-cps1
to the A board's reset signal to allow openkey to control it.  In
general you shouldn't need this, but waited to make it an available
option.  More details are in the README.md

The updated firmware is backwards compatible with the previous
hardware revision.